defer github API creation to PR handlers
Whenever a GitHub context is created, tokens are checked and refreshed.
Because GitHub tokens were being used across multiple jobs in a
repository-queue, it could happen that the tokens were expired once a
job later in the queue was started.

This change makes sure the GitHub API is created when the jobs is going
to be processed. That way the tokens are checked at the beginning of
each jobs, instead of at the beginning of the job queue.