Fix merging pull requests from forks
Before this change merging pull requests from forks
is broken because the branches array in the status
event was empty and the graphql query did return nothing.

This pr proposes using the sha field from the status
event and finding this sha hash  in all open PRs.
The commit is assigned to the PR if it's the leading
commit of this branch. (headRefOid).
It also simplifies the graphql query and add the
master restriction to the query.

The only assumption I took is that we either find the PR in
the 100 last updated PRs. If assumption is unreasonable
we would need to iterate over the whole PR set.